+++ src/clj/r99c/routes/home.clj
@@ -65,24 +65,28 @@
   "syntax check `answer`"
   [answer]
   (let [r (exec/sh ["gcc" "-xc" "-fsyntax-only" "-"] {:in answer})]
-    (zero? (:exit @r))))
+    (timbre/debug "validate" @r)
+    (:err @r)))
 
 (defn create-answer!
   "insert answer into answers table, compare the md5 value
    with other answers."
   [{:as request {:keys [num answer]} :params}]
   (if-let [error (validate-answer answer)]
-    (-> (redirect (str "/answer/" num))
-        (assoc :flash {:erros "syntax error"}))
+    (do
+     (timbre/debug "error" error)
+     (-> (redirect (str "/answer/" num))
+         (assoc :flash {:erros "syntax error"})))
     (let [login (name (get-in request [:session :identity]))
           ;; \n matches to \s
           stripped (-> (str/replace answer #"[ \t]" "")
                        remove-comments)
           md5 (digest/md5 stripped)]
-      (db/create-answer! {:login login}
-                      :num (Integer/parseInt num)
-                      :answer answer
-                      :md5 md5)
+      (timbre/debug "create-answer!")
+      (db/create-answer! {:login login
+                          :num (Integer/parseInt num)
+                          :answer answer
+                          :md5 md5})
       (redirect "/"))))
 
 (defn comment-form