+++ CHANGELOG.md
@@ -13,11 +13,14 @@ int func_test(void) {
 }
 ```
 * パスワードを変えられないと。
-* syntax check --- validation で？
+
 * 問題修正 ... test コラムを有効に。
 * /admin/users ... ユーザを一覧表示し、パスワードを初期化、エントリーを削除する。
 * /admin/comments ... 何をする予定だったっけ？
 
+## 0.6.0 - 2021-10-11
+### Added
+* syntax check, but not flash back.
 
 ## 0.5.1 - 2021-10-11
 ### Added